% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc1k}
\alias{pbmc1k}
\title{pbmc1k}
\format{
\subsection{pbmc1k}{

An object of class Seurat with 16053 features across 1037 samples within 2 assays:
\describe{
\item{Active assay:}{integrated (2000 features, 2000 variable features)}
\item{Other assays:}{RNA}
\item{Dimensional reductions calculated:}{pca, umap}
...
}
}
}
\source{
\url{https://satijalab.org/seurat/archive/v3.0/immune_alignment.html}
}
\usage{
pbmc1k
}
\description{
A Seurat object containing 1037 PBMCs from a Seurat tutorial.
}
\keyword{datasets}
