% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reference_Matrix_Builder.R
\name{Reference_Matrix_Builder}
\alias{Reference_Matrix_Builder}
\title{Reference_Matrix_Builder}
\usage{
Reference_Matrix_Builder(
  seurat_object,
  ident.1 = NULL,
  ident.2 = NULL,
  double.ident = TRUE,
  sep.double.ident = "_",
  reverse.double.ident = FALSE,
  clusters.1 = NULL,
  clusters.1.invert = FALSE,
  clusters.2 = NULL,
  clusters.2.invert = FALSE,
  downsample.object.first = NULL,
  downsample.object.last = NULL,
  downsample.cluster = NULL,
  automatic.downsample = FALSE,
  max.matrix.size = 900,
  file.name = "Reference_Matrix.txt",
  path = NULL,
  write.table = FALSE
)
}
\arguments{
\item{seurat_object}{A Seurat object.}

\item{ident.1}{Character. The first identity to use (for example "seurat_clusters" or "orig.ident"). Leave NULL if you want to use current identity (active.ident).}

\item{ident.2}{Character. The second identity to use. This will subset each ident.1 cluster with ident.2 identities (for example if ident.1 is "seurat_clusters" and ident.2 "orig.ident" cluster 0 will be divided into several clusters 0 each corresponding to an orig.ident). Leave NULL if you don't want to use a second identity.}

\item{double.ident}{Logical. If TRUE, each cell barcode will be renamed with a double identity (ident.1_ident.2). If FALSE, each cell barcode will be renamed with ident.1. Ignored if ident.2 is NULL.}

\item{sep.double.ident}{Character. The separator to use between ident.1 and ident.2 if double.ident = TRUE. Ignored if ident.2 is NULL.}

\item{reverse.double.ident}{Logical. If TRUE, the order of ident.1 and ident.2 will be reversed when creating the double identity (ident.2_ident.1). Ignored if ident.2 is NULL.}

\item{clusters.1}{Character. The clusters to subset from ident.1. Leave NULL if you want all clusters.}

\item{clusters.1.invert}{Logical. If TRUE, invert the selection from clusters.1 (if clusters.1 = c("0","1","2") and clusters.1.invert = TRUE, all clusters except 0, 1 and 2 will be selected).}

\item{clusters.2}{Character. The clusters to subset from ident.2. Leave NULL if you want all clusters.}

\item{clusters.2.invert}{Logical. If TRUE, invert the selection from clusters.2.}

\item{downsample.object.first}{Numeric. The number of cells to downsample from the Seurat object before subsetting clusters.1 and clusters.2. Leave NULL if you don't want to downsample.}

\item{downsample.object.last}{Numeric. The number of cells to downsample from the Seurat object after subsetting clusters.1 and clusters.2. Leave NULL if you don't want to downsample.}

\item{downsample.cluster}{Numeric. The number of cells to downsample from each cluster of ident.1. Will be performed after downsample.object.last. Leave NULL if you don't want to downsample.}

\item{automatic.downsample}{Logical. If TRUE, automatically downsample the Seurat object so that the Reference Matrix file written to disk would be just under the max.matrix.size limit (empirical). Performed after all other subset and downsample parameters. Ignored if write.table = FALSE. Please report an issue if you see a significant difference between the .txt or .tsv file written to disk and max.matrix.size (for example max.matrix.size is set to 200 MB but the .txt file is 400 MB or 100 MB).}

\item{max.matrix.size}{Numeric. The maximum size of the Reference Matrix file written to disk in MB. Will stop the function if the Reference Matrix file written to disk is estimated to be over this limit. Ignored if write.table = FALSE.}

\item{file.name}{Character. The name of the Reference Matrix file written to disk. Must not contain any space. You can specify .txt or .tsv for the output file.}

\item{path}{Character. The path to write the Reference Matrix into. Leave NULL for current working directory.}

\item{write.table}{Logical. If TRUE, write to disk the Reference Matrix as a .txt or .tsv file. If FALSE, only returns the Reference Matrix as a data.frame.}
}
\value{
A data.frame containing the Seurat object's RNA counts, with cell identities as column names and feature names as row names. If write.table = TRUE, a .txt or .tsv file of the data.frame is also written to disk.
}
\description{
This function builds a Reference Matrix for CIBERSORTx from a Seurat object.
}
\examples{
Reference_Matrix_Builder(
seurat_object = pbmc1k,
ident.1 = "seurat_clusters",
clusters.1 = c("Cluster.0","Cluster.4","Cluster.5"),
file.name = "Reference_Matrix_1.txt",
write.table = TRUE
)

Reference_Matrix_Builder(
seurat_object = pbmc1k,
ident.2 = "seurat_clusters",
downsample.object.first = 300,
write.table = TRUE
)

Reference_Matrix_Builder(
seurat_object = pbmc1k,
ident.1 = "seurat_clusters",
ident.2 = "orig.ident",
double.ident = FALSE,
clusters.1 = c("Cluster.7","Cluster.14"),
clusters.1.invert = TRUE,
clusters.2 = "IMMUNE_CTRL",
clusters.2.invert = TRUE,
file.name = "STIM_Reference_Matrix.tsv",
write.table = TRUE
)

}
