% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reference_Matrix_Sizer.R
\name{Reference_Matrix_Sizer}
\alias{Reference_Matrix_Sizer}
\title{Reference_Matrix_Sizer}
\usage{
Reference_Matrix_Sizer(
  seurat_object,
  max.matrix.size = 900,
  downsample = FALSE
)
}
\arguments{
\item{seurat_object}{A Seurat object.}

\item{max.matrix.size}{Numeric. The maximum size of the Reference Matrix file written to disk in MB.}

\item{downsample}{Logical. If TRUE, downsample the Seurat object so that the Reference Matrix file written to disk would be just under the max.matrix.size limit (empirical). Please report an issue if you see a significant difference between the .txt or .tsv file written to disk and max.matrix.size (for example max.matrix.size is set to 200 MB but the .txt file is 400 MB or 100 MB).}
}
\value{
The Seurat object, an estimation of the Reference Matrix size in MB and a suggestion of the number of cells to downsample. If downsample = TRUE, the Seurat object will instead be downsampled to the suggested number of cells.
}
\description{
This function calculates the size of a Reference Matrix which would be built from a Seurat object and suggests a number of cells to downsample from the object or performs that downsample to be under a specified size limit.
}
\examples{
Reference_Matrix_Sizer(
seurat_object = pbmc1k,
max.matrix.size = 20,
downsample = FALSE
)

Reference_Matrix_Sizer(
seurat_object = pbmc1k,
max.matrix.size = 50,
downsample = FALSE
)

}
